requirements.txt
qdrant-client
sentence-transformers
openai
rich
rich-theme-manager
pandas
python-dotenv
os
